# [PodSecurityPolicy] Optimize authorization check