@tallclair 2017-11-15
OK, I updated this to filter the validation errors by authorization on failure. It means that the failure case will be almost as slow as before (unless the pod is valid under most of the unauthorized profiles), but shouldn't affect the success case.

PTAL