@tallclair 2017-11-15
Rebased & update bazel. Updates since the initial review are in the second commit.