@tallclair 2017-11-20
`NewPlugin` wasn't used anywhere outside the file it's declared in, so I made it private. The matcher function doesn't really serve a purpose anymore, so I removed it (inlined the call to list).

PTAL.