@tallclair 2017-11-21
Squashed & rebased.