@ericchiang 2017-11-22
Kops job errored due to setup.py download failures

```
I1122 20:04:13.219]   Running setup.py (path:/tmp/pip-build-I38LYJ/PyYAML/setup.py) egg_info for package PyYAML
I1122 20:04:13.333]     
I1122 20:04:13.368] Downloading/unpacking docutils>=0.10 (from awscli)
I1122 20:04:13.504] Downloading/unpacking botocore==1.8.0 (from awscli)
I1122 20:04:13.874] Cleaning up...
I1122 20:04:14.302] Exception:
I1122 20:04:14.302] Traceback (most recent call last):
I1122 20:04:14.302]   File "/usr/lib/python2.7/dist-packages/pip/basecommand.py", line 122, in main
I1122 20:04:14.302]     status = self.run(options, args)
I1122 20:04:14.303]   File "/usr/lib/python2.7/dist-packages/pip/commands/install.py", line 290, in run
I1122 20:04:14.303]     requirement_set.prepare_files(finder, force_root_egg_info=self.bundle, bundle=self.bundle)
I1122 20:04:14.303]   File "/usr/lib/python2.7/dist-packages/pip/req.py", line 1198, in prepare_files
I1122 20:04:14.303]     do_download,
I1122 20:04:14.303]   File "/usr/lib/python2.7/dist-packages/pip/req.py", line 1376, in unpack_url
I1122 20:04:14.303]     self.session,
I1122 20:04:14.304]   File "/usr/lib/python2.7/dist-packages/pip/download.py", line 546, in unpack_http_url
I1122 20:04:14.304]     resp = session.get(target_url, stream=True)
I1122 20:04:14.304]   File "/usr/lib/python2.7/dist-packages/requests/sessions.py", line 469, in get
I1122 20:04:14.304]     return self.request('GET', url, **kwargs)
I1122 20:04:14.304]   File "/usr/lib/python2.7/dist-packages/pip/download.py", line 237, in request
I1122 20:04:14.304]     return super(PipSession, self).request(method, url, *args, **kwargs)
I1122 20:04:14.304]   File "/usr/lib/python2.7/dist-packages/requests/sessions.py", line 457, in request
I1122 20:04:14.305]     resp = self.send(prep, **send_kwargs)
I1122 20:04:14.305]   File "/usr/lib/python2.7/dist-packages/requests/sessions.py", line 569, in send
I1122 20:04:14.305]     r = adapter.send(request, **kwargs)
I1122 20:04:14.305]   File "/usr/lib/python2.7/dist-packages/requests/adapters.py", line 407, in send
I1122 20:04:14.305]     raise ConnectionError(err, request=request)
I1122 20:04:14.305] ConnectionError: ('Connection aborted.', error(104, 'Connection reset by peer'))
I1122 20:04:14.306] 
```

/test pull-kubernetes-e2e-kops-aws